video: tegra: host: Fix timeout race condition
This patch fixes a potential race condition in cases where the
timeout handler is executed at the same time with the submit
completed handler.

If the cdma lock was held by the submit completed handler the timeout
handler could not proceed until the submit completed handler was
finished. The submit completed handler changes the information about
the active work causing the timeout handler to observe information
that was not even related to the work causing the timeout.

Change-Id: Ia80706214f2fceb816c187caebecc8c6a47eb3cb
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/187481
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>