sync: Print sync dumps in debug level
Fence time outs are not usually error conditions; Fences may time out
as a result of lowering peripheral frequency too low to finish work
in time. In most cases the debug dump decreases signal-to-noise ratio
and hence it is not useful to print it out by default.

This patch makes the sync dumps to happen in debug level.

Change-Id: Idd09defac4778ccd3e4296577c5a186e0353d280
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/354167
GVS: Gerrit_Virtual_Submit
Reviewed-by: Timo Alho <talho@nvidia.com>
Reviewed-by: Shridhar Rasal <srasal@nvidia.com>
Reviewed-by: Jon Mayo <jmayo@nvidia.com>
Reviewed-by: Juha Tukkinen <jtukkinen@nvidia.com>