video: tegra: dc: Clear window option before window update.
There exits an issue that if window number is less than DC_N_WINDOWS,
window option of some windows won't be cleared. So although it should
be disabled, it might not be disabled properly. This will lead to the
failure of scan-out on screen.

Bug 943846


from: aosp kernel/tegra
commit: 90b79e5712300baab889772a5af348559ac95836