video: tegra: dc: fix MC_DECERR display0_win errors on shutdown
When a DC window is being released, a NULL flip is used to indicate
that the window should be disabled.  To disable a window, 0 is
written to WIN_OPTIONS.

The MC_DECERR on window shutdown is a symptom of forgetting step 2)
below when disabling windows, leaving no indication that there is
anything to wait for.  This causes DC to erroneously unpin the
scanout buffer while the buffer is still actively being used.

Summary of flip (window update) synchronization in DC:
1) program some window registers
2) set win->dirty=1
3) schedule the activation of the registers
4) request VBLANK or HBLANK interrupts
5) wait for win->dirty==0

In the trigger_windows ISR (every VBLANK or HBLANK):
1) if there is no window update pending in HW, clear dirty flag to
   indicate that ACTIVE registers are up-to-date.