video: tegra: host: Dump sync queue in debug dump
Dump contents of sync queue when debug dump is requested, either
because of a stuck sync point, or when the debugfs entry is accessed.

Exclude FIFO from normal debugfs dump. Dumping FIFO is invasive and
actually changes how the channel behaves. It's safe to dump only when
system is in fault condition.