video: tegra: host: use constant EMC clock for MPE
MPE clock is tied to MPE EMC clock. Although MPE clock
goes low due to DFS, lowering MPE EMC is not recommended
because video recording is high memory bandwidth use case.
MPE EMC is set to 400MHz initially and won't change even
if MPE is set to a new clock frequency.

For now, 400MHz is used for MPE EMC. But later it might be
tuned in a way to have relatioship with MPE clock.
For example, MPE EMC could be (MPE + 100MHz). To find
a right value, we need more experiment.