video: tegra: hdcp: Fix Locking Issue
Locking in nvhdcp was set up in such a way that we couldn't
change the state of the HDMI plug while negotiating the HDCP connection.

Another lock was added that is only taken when the state of the plug or
hdcp session are changed.  This enables correct failure if the plug is removed
during negotiation.

Bug 1024031