drivers: misc: publish fps, cpu load, thread count
adapted cpufreq_interactive load estimation and averaging of the
number of runnable threads to a misc device that will write the
current cpu id, load and nr_runnable_threads for each cpu to a sysfs
node. Intended to provide an accurate cpu load reading to user space
scaling controllers. In addition the momentary frame rate is written
to /d/fps.

Bug 1161410
Bug 1164121