misc: tegra-throughput: Use notifiers
We used to deliver this information by calling directly a function
inside nvhost_pod. This is not suitable for our needs as we want
to be able to use the same governor with several different (graphics)
devices in the same system.

This patch modifies tegra-throughput to use notifier chain for
delivering information about changed throughput hint.

Bug 1364804