cpufreq: tegra2 voltage control: Improve undervolt table
__tegra_dvfs_set_rate() in dvfs.c searches for the correct voltage index
to apply to a clock frequency.

Use this same format for the user undervolt table. Do the extra work
in cpufreq sysfs interfaces to map the user voltages to the appropriate
array index.

This way the CPU_DVFS definitions can be restored to their stock values.
And voltage control is compatible any CPU_DVFS definition.

Signed-off-by: RyanG <decatf@gmail.com>